cureCuma/services/back/requirements.txt
flask_bcrypt
requests
SQLAlchemy
pymysql
pytest
IPython

gunicorn
python-dateutil

pandas